machine-learning/pyproject.toml
[project]
name = "iis-projekt-energija"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "dvc[s3]>=3.67.0",
    "numpy>=1.22.4,<2.0.0",
    "pandas>=2.2.0,<3.0.0",
    "pyaml>=26.2.1",
    "requests>=2.33.0",
    "scikit-learn>=1.8.0",
    "great-expectations==0.18.21",
    "evidently>=0.7.1,<0.8.0",
    "tensorflow>=2.16.0,<2.17.0",
    "matplotlib>=3.10.9",
    "mlflow>=2.22.0,<3.0.0",
    "python-dotenv>=1.2.2",
]
